refactor(server): remove unused imports
